# Warn about gauge re-registration