--- a/benchmarks/benchmarks-core/src/jmh/resources/logback.xml
+++ b/benchmarks/benchmarks-core/src/jmh/resources/logback.xml
@@ -1,0 +1,26 @@
+<!--
+	Copyright 2024 VMware, Inc.
+
+	Licensed under the Apache License, Version 2.0 (the "License");
+	you may not use this file except in compliance with the License.
+	You may obtain a copy of the License at
+
+	https://www.apache.org/licenses/LICENSE-2.0
+
+	Unless required by applicable law or agreed to in writing, software
+	distributed under the License is distributed on an "AS IS" BASIS,
+	WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+	See the License for the specific language governing permissions and
+	limitations under the License.
+-->
+<configuration>
+<appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
+	<encoder>
+		<pattern>%d{HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n</pattern>
+	</encoder>
+</appender>
+
+<root level="warn">
+	<appender-ref ref="STDOUT" />
+</root>
+</configuration>